pyproject.toml
[project]
name = "django-base"
version = "0.1.0"
description = "Base para um projeto Django / Base for a Django project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-prometheus>=2.4.1",
    "django-q2>=1.8.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]

[tool.setuptools]
packages = { find = { where = ["src"] } }

# Ruff Linter and Formatter Configuration
# Configuração do Linter e Formatador Ruff
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4"]
ignore = []

# Coverage.py Configuration
# Configuração do Coverage.py
[tool.coverage.run]
source = ["src/core"]
omit = [
    "src/*/migrations/*",
    "src/*/tests/*",
    "src/*/admin.py",
    "src/django_base/*",
    "manage.py",
]

[tool.coverage.report]
show_missing = true

[dependency-groups]
dev = [
    "coverage>=7.10.7",
    "django-extensions>=4.1",
    "pre-commit>=4.3.0",
    "ruff>=0.13.2",
    "watchdog>=6.0.0",
    "werkzeug>=3.1.3",
]

